added missing namespaces
