+++ Mico Emotion/Assets/Main/Scripts/Screen1/MovingCloud.cs
@@ -2,28 +2,31 @@
 
 using DG.Tweening;
 
-public class MovingCloud : MonoBehaviour
+namespace Emotion.Screen1
 {
-    #region FIELDS
-
-    private const float MinSpeed = 25.0f;
-    private const float MaxSpeed = 40.0f;
-    private const float MaxXPosition = 25.0f;
-
-    #endregion
-
-    #region BEHAVIORS
-
-    private void Awake()
+    public class MovingCloud : MonoBehaviour
     {
-        Repeat();
-    }
+        #region FIELDS
 
-    private void Repeat()
-    {
-        float startValue = Random.value > 0.5 ? 1.0f : -1.0f;
-        transform.DOMoveX(-MaxXPosition * startValue, Random.Range(MinSpeed, MaxSpeed)).From(MaxXPosition * startValue).OnComplete(Repeat);
-    }
+        private const float MinSpeed = 25.0f;
+        private const float MaxSpeed = 40.0f;
+        private const float MaxXPosition = 25.0f;
 
-    #endregion
+        #endregion
+
+        #region BEHAVIORS
+
+        private void Awake()
+        {
+            Repeat();
+        }
+
+        private void Repeat()
+        {
+            float startValue = Random.value > 0.5 ? 1.0f : -1.0f;
+            transform.DOMoveX(-MaxXPosition * startValue, Random.Range(MinSpeed, MaxSpeed)).From(MaxXPosition * startValue).OnComplete(Repeat);
+        }
+
+        #endregion
+    }
 }
