+++ Mico Emotion/Assets/Main/Utilities/Gestures.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a1f41e41fc4d342b38462e7b53792289
+guid: 1b20ec775362a4c78ab2eb2c37383c2d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
