+++ Mico Emotion/Assets/Main/Utilities/Gestures/DoubleClick.cs.meta
